+++ MarioGame/Brick.h
@@ -8,5 +8,5 @@ class CBrick : public CGameObject
 {
 public:
 	virtual void Render();
-	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
+	virtual void GetBoundingBox(double& l, double& t, double& r, double& b);
 };
