+++ MarioGame/GameObject.cpp
@@ -25,12 +25,12 @@ void CGameObject::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 
 void CGameObject::RenderBoundingBox()
 {
-	D3DXVECTOR3 p(x, y, 0);
+	D3DXVECTOR3 p((float) x, (float) y, 0);
 	RECT rect;
 
 	LPDIRECT3DTEXTURE9 bbox = CTextures::GetInstance()->Get(ID_TEX_BBOX);
 
-	float l, t, r, b;
+	double l, t, r, b;
 
 	GetBoundingBox(l, t, r, b);
 	rect.left = 0;
